requirements.txt
google-generativeai
datasets
huggingface_hub
tqdm
pyarrow